Fix TCODEID-2s loading problem
Since obs is sourced from templater, which loads late, this code does
not have the necessary apis to run on start, which breaks any first note.
So on cold startups , i have it fail gracefully, then rerun the code after
20seconds (the avg time it takes fora an obsidian plugin cold start).

I also suffixed the file with a shorthand ID for easy reference in commit messages.